requirements.txt
tensorflow-gpu
numpy>=1.18.5
sklearn
matplotlib>=3.2.2
pandas>=1.2.1